requirements.txt
APScheduler==3.11.0
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
colorama==0.4.6
cryptography==44.0.3
cssselect2==0.8.0
defusedxml==0.7.1
dj-database-url==3.0.1
Django==5.2
django-apscheduler==0.7.0
django-cors-headers==4.7.0
django-extensions==4.1
django-heroku==0.3.1
djangorestframework==3.16.0
docutils==0.21.2
gunicorn==23.0.0
html5lib==1.1
idna==3.10
kivy-deps.angle==0.4.0
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.7.0
lxml==5.4.0
optional-django==0.3.0
oscrypto==1.3.0
packaging==25.0
pillow==11.2.1
psycopg2==2.9.11
pycparser==2.22
Pygments==2.18.0
pyHanko==0.26.0
pyhanko-certvalidator==0.26.8
pypdf==5.4.0
pypiwin32==223
python-bidi==0.6.6
pywin32==308
PyYAML==6.0.2
qrcode==8.2
reportlab==4.4.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.2.3
webencodings==0.5.1
whitenoise==6.11.0
xhtml2pdf==0.2.17
